Replacement needs to remove suspected node from active view.
When replacing a node, the suspected node must be removed from the active view to ensure there is room for the replacement node.  Similarly, if that node was holding a reserved slot, it has to give up that slot.  Ensure replaced node moves into the passive view and the active view is populated.