Stratify calls from accepting messages to reduce nondeterminism.
In some cases, functions, when invoked, will take time to execute, which means that even if messages are delivered in order, messages might be delivered before/during/or after the call in various interleavings.

This causes issues when performing message replay and shrinking because of the variation in excecution time because of the preemptive scheduler of Erlang.

For instance, if I want to make an assertion while 5 messages are being delivered and it takes time for me to perform the assertion, I may see any number of messages 1 - 5.  Therefore, ensure that before we enter a function boundary, we have everything we should have, prevent delivery during execution of a function, and then wait for delivery before proceeding on.

This is not done during schedule exploration, this is done only when replaying a traced schedule or shrinking a scheduler to produce a minimal counterexample.