Prevent race condition.
Prevent a race where we want to crash a node and we remove the partition beforehand -- in the split second between the partition removal and crash insertion, we can trigger an issue where a message that's supposed to be dropped ends up begin delivered.  Crash nodes with partition intact - this won't work if we restart the nodes, but in our current fault-injection strategy, we do not recover from the final crash fault.