Restrict schedule generation.
When using a schedule that contains redundancy in it -- if we attempt to run this schedule without omissions the schedule will contain messages that will never come: this is because redundancy is only visible under omissions.  Therefore, when we append more schedules into the next analysis pass, we need to supply a base set of omissions that are assumed for the next run.