Rewrite scheduler.
Rewrite the scheduler so that when redundancy is identified, we use the schedule that generated the redundancy and then explore failures from extensions of that schedule.  Attempt to prune schedules as soon as we have enough information to, to reduce the number we need to both materialize and perform schedule exploration on.