Basic version of remote monitoring
Given a {partisan_remote_reference, Node, {partisan_process_reference,
PidAsList}}, you will be able to call partisan_monitor:monitor/1 in order to
monitor the process. This will message a named process on the Node to begin
monitoring the PidAsList as a proxy for the calling process, returning a
partisan_encoded_reference.

If the monitored process dies, it will send the down message to the local
partisan_monitor process which will do a lookup and forward the message off to
the remote monitoring process.

You can also call partisan_monitor:demonitor/1, handing it the
partisan_encoded_reference, in order to stop monitoring the remote process.