fix implementation of plumtree heartbet using interval sets
The previous solution would accummulate an infinity amount of timestamps with form

```
{node(), erlang:unique_integer([monotonic, positive]}.
```

We replaced a new timestamps that contains node(), timestamp and an inversal set of monotonic counters

```
{'node2@127.0.0.1',1696367521778485625,[{1,2}]}
```

This ensure we have at most a single record per node. The interval set allows to compress the set i.e. if a node1 received all first 1M messages from node2 the entry would be

```
{'node2@127.0.0.1',1696367521778485625,[{1,1000001}]}
```

Signed-off-by: Alejandro M. Ramallo <alejandro.ramallo@leapsight.com>